+++ background.h
@@ -7,10 +7,8 @@
 #include "dependencies/thpool.h"        // thread pool
 
 #define THREAD_MOO_VERSION      "2.5"   // Version of our MOO threading library.
-#define TOTAL_BACKGROUND_THREADS 2      // The total number of background threads running in the pool.
 #define MAX_BACKGROUND_THREADS  20      /* The total number threads allowed to be run from within the MOO.
                                            Can be overridden with $server_options.max_background_threads */
-#define DEFAULT_THREAD_MODE     true    // The default behavior of threaded functions without a call to set_thread_mode
 
 typedef struct background_waiter {
     vm the_vm;                          // Where we resume when we're done.